Refactored FamilyListRepository, added IFamilyListRepository

+++ harmonee/Server/Data/FamilyListRepository.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace harmonee.Server.Data
 {
-    public class FamilyListRepository : IRepository<FamilyList>
+    public class FamilyListRepository : IFamilyListRepository
     {
         private readonly FamilyListContext _context;
 
@@ -12,33 +12,61 @@ namespace harmonee.Server.Data
             _context = context;
         }
 
-        public FamilyList Add(FamilyList entity)
+        public bool Add(FamilyList entity)
         {
+            var existingList = _context.FamilyLists.FirstOrDefault(fl => fl.Id == entity.Id);
+            if (existingList is not null)
+            {
+                return false;
+            }
+
             _context.FamilyLists.Add(entity);
             _context.SaveChanges();
-            return entity;
+            return true;
         }
 
-        public IEnumerable<FamilyList> AddMany(IEnumerable<FamilyList> entities)
+        public bool AddMany(IEnumerable<FamilyList> entities)
         {
-            foreach (var family in entities)
+            var contextChanged = false;
+            var incomingIds = entities.Select(fl => fl.Id).ToHashSet();
+            var existingFamilyListIds = _context.FamilyLists.Where(fl => incomingIds.Contains(fl.Id)).Select(fl => fl.Id).ToHashSet();
+            var listsToAdd = entities.Where(fl => !existingFamilyListIds.Contains(fl.Id));
+
+            foreach (var listToAdd in listsToAdd)
             {
-                _context.FamilyLists.Add(family);
+                _context.Add(listToAdd);
+                contextChanged = true;
             }
-            _context.SaveChanges();
-            return entities;
+
+            if (contextChanged)
+            {
+                _context.SaveChanges();
+            }
+
+            return contextChanged;
         }
 
         public bool Delete(Guid id)
         {
-            _context.FamilyLists.Remove(GetById(id));
-            _context.SaveChanges();
-            return true;
+            var existingFamilyList = _context.FamilyLists.FirstOrDefault(fl => fl.Id == id);
+            if (existingFamilyList is not null)
+            {
+                _context.FamilyLists.Remove(GetById(id));
+                _context.SaveChanges();
+                return true;    
+            }
+
+            return false;
         }
 
         public bool DeleteMany(IEnumerable<Guid> ids)
         {
             var familiesToRemove = _context.FamilyLists.Where(f => ids.Contains(f.Id));
+            if (familiesToRemove.Count() == 0)
+            {
+                return false;
+            }
+
             _context.FamilyLists.RemoveRange(familiesToRemove);
             _context.SaveChanges();
             return true;
@@ -61,6 +89,17 @@ namespace harmonee.Server.Data
 
         public bool Update(FamilyList entity)
         {
+            var existingList = _context.FamilyLists.FirstOrDefault(fl => fl.Id == entity.Id);
+            if (existingList is null)
+            {
+                return false;
+            }
+            
+            if (entity.Equals(existingList))
+            {
+                return false;
+            }
+
             _context.FamilyLists.Update(entity);
             _context.SaveChanges();
             return true;
@@ -68,9 +107,25 @@ namespace harmonee.Server.Data
 
         public bool UpdateMany(IEnumerable<FamilyList> entities)
         {
-            _context.FamilyLists.UpdateRange(entities);
-            _context.SaveChanges();
-            return true;
+            var contextChanged = false;
+            var incomingListIds = entities.Select(fl => fl.Id).ToHashSet();
+            var existingRecords = _context.FamilyLists.Where(fl => incomingListIds.Contains(fl.Id));
+            foreach (var existingRecord in existingRecords)
+            {
+                var incomingRecord = entities.FirstOrDefault(fl => fl.Id == existingRecord.Id);
+                if (!incomingRecord.Equals(existingRecord))
+                {
+                    _context.Update(incomingRecord);
+                    contextChanged = true;
+                }
+            }
+
+            if (contextChanged)
+            {
+                _context.SaveChanges();
+            }
+
+            return contextChanged;
         }
     }
 
@@ -86,8 +141,8 @@ namespace harmonee.Server.Data
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.UseSqlServer(_connectionString);
             // Use appropriate server and connection string
+            optionsBuilder.UseSqlServer(_connectionString);
         }
     }
 }
